+++ Strata/StrataCore/Models/TimelineEvent.swift
@@ -27,6 +27,7 @@ public enum TimelineSource: String, CaseIterable, Sendable, Codable {
     case usn        = "USN"
     case srum       = "SRUM"
     case browser    = "BROWSER"
+    case mft        = "MFT"
 
     public var label: String {
         switch self {
@@ -35,6 +36,7 @@ public enum TimelineSource: String, CaseIterable, Sendable, Codable {
         case .usn:        return "USN Journal"
         case .srum:       return "SRUM"
         case .browser:    return "Browser History"
+        case .mft:        return "MFT"
         }
     }
 }